+++ CHANGELOG.md
@@ -4,7 +4,12 @@
 Version 4.1-SNAPSHOT
 -------------
 
-Version 4
+ADDED:
+
+- new methods in the class Argument for cover the checks for the primitive array types
+- new methods in the class Check for cover the checks for the primitive array types
+  
+  Version 4
 -------------
 
 ADDED: